# Conflicts:
#	src/infra/email/contexts.py
#	src/services/auth_service.py
#	src/utils.py

+++ src/services/auth_service.py
@@ -9,13 +9,15 @@ from starlette.status import (
     HTTP_409_CONFLICT,
     HTTP_400_BAD_REQUEST,
     HTTP_401_UNAUTHORIZED,
+    HTTP_500_INTERNAL_SERVER_ERROR,
 )
+
 from jwt.exceptions import ExpiredSignatureError, DecodeError
 
 from src.repositories.general_user_repo import UserRepository
 from src.models.app_models import User, VerificationCode
 from src.db import get_db_session
-from src.repositories.intern_repository import InternRepository
+from src.repositories.intern_repo import InternRepository
 from src.repositories.verification_code_repo import VerificationCodeRepository
 from src.schemas import UserInModel, InternInModel
 from src.schemas.user_schemas import UserOutModel
@@ -24,7 +26,7 @@ from src.utils import (
     password_is_correct,
     hash_password,
     generate_random_code,
-    normalize_email,
+    normalize_string,
     generate_password_reset_token,
     decode_token,
     TokenType,
@@ -37,6 +39,10 @@ from ..infra.email.contexts import (
 )
 from ..infra.email import send_email
 from ..logger import logger
+from ..repositories import SkillRepository
+from ..repositories.supervisor_repo import SupervisorRepository
+from ..schemas.intern_schemas import InternOutModel
+from ..schemas.supervisor_schemas import SupervisorInModel, SupervisorOutModel
 
 
 class AuthService:
@@ -45,18 +51,22 @@ class AuthService:
         session: Annotated[AsyncSession, Depends(get_db_session)],
         user_repo: Annotated[UserRepository, Depends()],
         intern_repo: Annotated[InternRepository, Depends()],
+        supervisor_repo: Annotated[SupervisorRepository, Depends()],
         code_repo: Annotated[VerificationCodeRepository, Depends()],
+        skill_repo: Annotated[SkillRepository, Depends()],
         background_task: BackgroundTasks,
     ):
         self.session = session
         self.user_repo = user_repo
         self.intern_repo = intern_repo
+        self.supervisor_repo = supervisor_repo
         self.code_repo = code_repo
+        self.skill_repo = skill_repo
         self.background_task = background_task
 
     # TODO: Rate limit
     async def create_unverified_new_user(
-        self, new_user: UserInModel | InternInModel
+        self, new_user: UserInModel | InternInModel | SupervisorInModel
     ) -> dict[str, str]:
         async with self.session.begin():  # Transactional, for atomicity
             existing_user: User = await self.user_repo.get_user_by_email_or_phone(
@@ -69,42 +79,44 @@ class AuthService:
                 if existing_user.verified:
                     raise HTTPException(
                         status_code=HTTP_409_CONFLICT,
-                        detail="User already exists. Please log in",
+                        detail="User already exists with specified email or phone number. Please log in",
                     )
 
+                # The user exists but isn't verified yet
                 code: str = generate_random_code()
                 await self.code_repo.upsert_code_with_user_id(
                     conn=self.session, user_id=existing_user.id, value=code
                 )
                 send_code = code
-                user_email = normalize_email(existing_user.email)
+                user_email = normalize_string(existing_user.email)
 
             else:
-                # TODO: refactor and simplify different user creation. To use Polymophic/Joined table inheritance
-
                 # Hash password  when creating a new user
                 new_user.password = hash_password(password=new_user.password)
 
-                unverified_user: User = (
-                    await self.intern_repo.create_new_intern(
-                        conn=self.session, new_intern=new_user
-                    )  # if the type is Intern
-                    if isinstance(new_user, InternInModel)
-                    else await self.user_repo.create_new_user(  # if type: any other type
-                        conn=self.session, new_user=new_user
-                    )
-                    # TODO expand this to accomodate HR (admin) entities
+                unverified_user: User = await self.user_repo.create_new_user(
+                    conn=self.session, new_user=new_user, skill_repo=self.skill_repo
                 )
+
+                if isinstance(new_user, InternInModel):
+                    unverified_user: User = await self.intern_repo.create_new_intern(
+                        conn=self.session, new_intern=new_user, user=unverified_user
+                    )
+                elif isinstance(new_user, SupervisorInModel):
+                    unverified_user: User = (
+                        await self.supervisor_repo.create_new_supervisor(
+                            conn=self.session,
+                            new_supervisor=new_user,
+                            user=unverified_user,
+                        )
+                    )
 
                 code = generate_random_code()
                 await self.code_repo.create_code(
                     conn=self.session, user_id=unverified_user.id, code=code
                 )
                 send_code = code
-                user_email = normalize_email(unverified_user.email)
-
-        # At this point, transaction has committed successfully
-        print(f"Code for {user_email}: {send_code}")
+                user_email = normalize_string(unverified_user.email)
 
         # Send verification code to email
         self.background_task.add_task(
@@ -130,14 +142,29 @@ class AuthService:
                 user_id=verification_code.user_id,
                 values={"verified": True},
             )
+            if not verified_user:
+                raise HTTPException(
+                    status_code=HTTP_500_INTERNAL_SERVER_ERROR,
+                    detail="Something went wrong with verification",
+                )
+
             await self.code_repo.delete_code(conn=self.session, value=code)
 
+            if verified_user.type == UserType.SUPERVISOR:
+                user_to_login = SupervisorOutModel.from_supervisor(verified_user)
+                logger.info(f"Supervisor user {verified_user.email} has been verified")
+            elif verified_user.type == UserType.INTERN:
+                user_to_login = InternOutModel.from_intern(verified_user)
+                logger.info(f"Intern user {verified_user.email} has been verified")
+            else:
+                user_to_login = UserOutModel.from_user(verified_user)
+
             access_token: str = generate_access_token(
-                UserOutModel.from_user(verified_user),
+                user_to_login=user_to_login,
             )
 
         # Send confirmation mail once user has been created.
-        user_normalized_email: str = normalize_email(verified_user.email)
+        user_normalized_email: str = normalize_string(verified_user.email)
         self.background_task.add_task(
             send_email,
             user_normalized_email,
@@ -164,9 +191,17 @@ class AuthService:
                 status_code=HTTP_401_UNAUTHORIZED, detail="Invalid login credentials"
             )
 
-        user_to_login: UserOutModel = UserOutModel.from_user(existing_user)
+        if existing_user.type == UserType.SUPERVISOR:
+            user_to_login: UserOutModel = SupervisorOutModel.from_supervisor(
+                existing_user
+            )
+        elif existing_user.type == UserType.INTERN:
+            user_to_login: UserOutModel = InternOutModel.from_intern(existing_user)
+        else:
+            user_to_login: UserOutModel = UserOutModel.from_user(existing_user)
 
         access_token: str = generate_access_token(user_to_login)
+
         # new_refresh_token: str = await generate_refresh_token(
         #     existing_user.email, self.token_repo
         # )
@@ -247,7 +282,7 @@ class AuthService:
         self.background_task.add_task(
             send_email,
             updated_user.email,
-            context=UpdatedUserContext(values_updated=values_to_update),
+            context=UpdatedUserContext(values_updated=list(values_to_update.keys())),
         )
         return {
             "detail": "Your password has been reset successfully, please proceed to log in"
